Added parser abstraction
