+++ Calculator/Calculator/Parser/IParser.cs
@@ -1,0 +1,7 @@
+﻿namespace Calculator.Parser
+{
+    public interface IParser
+    {
+        Expression Parse(string input);
+    }
+}
